% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_form_response.R
\name{parse_form_response}
\alias{parse_form_response}
\title{Парсер и форматирование результатов google формы}
\usage{
parse_form_response(
  responses_dox_id,
  report_dox_id = responses_dox_id,
  response_sheet_mask = "Form Responses|Ответы на форму",
  service_key_path = system.file("service.json", package = "gformparser")
)
}
\arguments{
\item{responses_dox_id}{ID Докса с результатами формы}

\item{report_dox_id}{ID Докса в который надо выплюнуть результаты}

\item{response_sheet_mask}{Маска имён листов на которых хранятся результаты}

\item{service_key_path}{Путь к файлу сервисного аккаунта}
}
\value{
Возвращает информацию о новом доксе.
}
\description{
Парсер и форматирование результатов google формы
}
